requirements.txt
mariadb
python-dotenv
fastapi[standard]
uvicorn[standard]
pydantic
